+++ ratarmount.py
@@ -250,6 +250,7 @@ class SQLiteIndexedTar:
         writeIndex                 = False,
         clearIndexCache            = False,
         indexFileName              = None,
+        indexFolders               = None,
         recursive                  = False,
         gzipSeekPointSpacing       = 4*1024*1024,
         encoding                   = tarfile.ENCODING,
@@ -325,11 +326,19 @@ class SQLiteIndexedTar:
                 pass
 
         # will be used for storing indexes if current path is read-only
-        possibleIndexFilePaths = [ os.path.abspath( indexFileName ) ] if indexFileName is not None else [
-            self.tarFileName + ".index.sqlite",
-            os.path.expanduser( os.path.join( "~", ".ratarmount",
-                                              self.tarFileName.replace( "/", "_" ) + ".index.sqlite" ) )
-        ]
+        possibleIndexFilePaths = [ self.tarFileName + ".index.sqlite" ]
+        indexPathAsName = self.tarFileName.replace( "/", "_" ) + ".index.sqlite"
+        if indexFileName:
+            # A given index file name takes precedence and there should be no implicit fallback
+            possibleIndexFilePaths = [ os.path.abspath( os.path.expanduser( indexFileName ) ) ]
+        elif indexFolders:
+            # An empty path is to be interpreted as the default path right besides the TAR
+            if '' not in indexFolders:
+                possibleIndexFilePaths = []
+            for folder in indexFolders:
+                if folder:
+                    indexPath = os.path.join( folder, indexPathAsName )
+                    possibleIndexFilePaths.append( os.path.abspath( os.path.expanduser( indexPath ) ) )
 
         self.indexFileName = None
         if clearIndexCache:
@@ -1699,36 +1708,49 @@ With ratarmount, you can:
 # Metadata Index Cache
 
 In order to reduce the mounting time, the created index for random access
-to files inside the tar will be saved to these locations in order. A lower
-location will only be used if all upper locations can't be written to.
+to files inside the tar will be saved to one of these locations. These
+locations are checked in order and the first, which works sufficiently, will
+be used. This is the default location order:
 
-    1. <path to tar>.index.sqlite
-    2. ~/.ratarmount/<path to tar: '/' -> '_'>.index.sqlite
-       E.g., ~/.ratarmount/_media_cdrom_programm.tar.index.sqlite
+  1. <path to tar>.index.sqlite
+  2. ~/.ratarmount/<path to tar: '/' -> '_'>.index.sqlite
+     E.g., ~/.ratarmount/_media_cdrom_programm.tar.index.sqlite
+
+This list of fallback folders can be overwritten using the `--index-folders`
+option. Furthermore, an explicitly named index file may be specified using
+the `--index-file` option. If `--index-file` is used, then the fallback
+folders, including the default ones, will be ignored!
 
 # Bind Mounting
 
 The mount sources can be TARs and/or folders.  Because of that, ratarmount
 can also be used to bind mount folders read-only to another path similar to
-"bindfs" and "mount --bind". So, for:
+`bindfs` and `mount --bind`. So, for:
+
     ratarmount folder mountpoint
-all files in folder will now be visible in mountpoint.
+
+all files in `folder` will now be visible in mountpoint.
 
 # Union Mounting
 
 If multiple mount sources are specified, the sources on the right side will be
 added to or update existing files from a mount source left of it. For example:
+
     ratarmount folder1 folder2 mountpoint
+
 will make both, the files from folder1 and folder2, visible in mountpoint.
 If a file exists in both multiple source, then the file from the rightmost
-mount source will be used, which in the above example would be "folder2".
+mount source will be used, which in the above example would be `folder2`.
 
 If you want to update / overwrite a folder with the contents of a given TAR,
 you can specify the folder both as a mount source and as the mount point:
+
     ratarmount folder file.tar folder
+
 The FUSE option -o nonempty will be automatically added if such a usage is
 detected. If you instead want to update a TAR with a folder, you only have to
 swap the two mount sources:
+
     ratarmount file.tar folder folder
 
 # File versions
@@ -1736,15 +1758,19 @@ swap the two mount sources:
 If a file exists multiple times in a TAR or in multiple mount sources, then
 the hidden versions can be accessed through special <file>.versions folders.
 For example, consider:
+
     ratarmount folder updated.tar mountpoint
-and the file "foo" exists both in the folder and in two different versions
-in "updated.tar". Then, you can list all three versions using:
+
+and the file `foo` exists both in the folder and as two different versions
+in `updated.tar`. Then, you can list all three versions using:
+
     ls -la mountpoint/foo.versions/
         dr-xr-xr-x 2 user group     0 Apr 25 21:41 .
         dr-x------ 2 user group 10240 Apr 26 15:59 ..
         -r-x------ 2 user group   123 Apr 25 21:41 1
         -r-x------ 2 user group   256 Apr 25 21:53 2
         -r-x------ 2 user group  1024 Apr 25 22:13 3
+
 In this example, the oldest version has only 123 bytes while the newest and
 by default shown version has 1024 bytes. So, in order to look at the oldest
 version, you can simply do:
@@ -1828,8 +1854,19 @@ version, you can simply do:
 
     parser.add_argument(
         '--index-file', type = str,
-        help = 'Specify a path to the .index.sqlite file.')
+        help = 'Specify a path to the .index.sqlite file. Setting this will disable fallback index folders.' )
 
+    parser.add_argument(
+        '--index-folders', default = "," + os.path.join( "~", ".ratarmount" ),
+        help = 'Specify one or multiple paths for storing .index.sqlite files. Paths will be tested for suitability '
+               'in the given order. An empty path will be interpreted as the location in which the TAR resides. '
+               'If the argument begins with a bracket "[", then it will be interpreted as a JSON-formatted list. '
+               'If the argument contains a comma ",", it will be interpreted as a comma-separated list of folders. '
+               'Else, the whole string will be interpreted as one folder path. Examples: '
+               '--index-folders ",~/.foo" will try to save besides the TAR and if that does not work, in ~/.foo. '
+               '--index-folders \'["~/.ratarmount", "foo,9000"]\' will never try to save besides the TAR. '
+               '--index-folder ~/.ratarmount will only test ~/.ratarmount as a storage location and nothing else. '
+               'Instead, it will first try ~/.ratarmount and the folder "foo,9000". ' )
 
     parser.add_argument(
         '-o', '--fuse', type = str, default = '',
@@ -1888,6 +1925,13 @@ version, you can simply do:
             args.mount_point = os.path.splitext( tarPath )[0]
     args.mount_point = os.path.abspath( args.mount_point )
 
+    # Preprocess the --index-folders list as a string argument
+    if args.index_folders:
+        if args.index_folders[0] == '[':
+            args.index_folders = json.loads( args.index_folders )
+        elif ',' in args.index_folders:
+            args.index_folders = args.index_folders.split( ',' )
+
     return args
 
 def cli( args = None ):
@@ -1922,6 +1966,7 @@ def cli( args = None ):
         verifyModificationTime     = args.verify_mtime,
         stripRecursiveTarExtension = args.strip_recursive_tar_extension,
         indexFileName              = args.index_file,
+        indexFolders               = args.index_folders,
     )
 
     fuse.FUSE( operations = fuseOperationsObject,